pyproject.toml
[project]
name = "phyclip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.4.0+cu124; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torch==2.4.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
    "torchvision==0.19.0+cu124; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torchvision==0.19.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
    "ftfy>=6.0",
    "hydra-core>=1.1",
    "loguru>=0.6",
    "nltk>=3.6",
    "tensorflow-datasets>=4.9.0",
    "torchmetrics>=1.0",
    "tqdm>=4.50.0",
    "webdataset>=0.2",
    "wordsegment>=1.3",
    "opencv-python>=4.9.0.80",
    "albumentations==1.4.16",
    "numpy==1.26.2",
    "tfds-nightly>=4.9.3.dev202311230044",
    "seaborn>=0.13.2",
    "timm>=1.0.15",
    "vl-checklist>=0.0.1",
    "open-clip-torch>=3.1.0",
    "pillow>=11.2.1",
    "omegaconf>=2.3.0",
    "fsspec>=2025.3.2",
    "networkx>=3.2.1",
    "pandas>=2.2.3",
    "huggingface-hub>=0.31.1",
    "matplotlib>=3.10.3",
    "regex>=2024.11.6",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "tensorboard>=2.20.0",
]

[tool.uv]
find-links = [
    "https://download.pytorch.org/whl/cu124/torch",
    "https://download.pytorch.org/whl/cu124/torchvision",
]

